TASK: added tests for DM, renamed packages
