+++ src/test/java/com/corevalue/constant/Tags.java
@@ -1,0 +1,10 @@
+package com.corevalue.constant;
+
+public enum Tags {
+    BUTTON, UL, LI, A, DIV;
+
+    @Override
+    public String toString() {
+        return this.name().toLowerCase();
+    }
+}
