--- src/test/java/com/corevalue/constant/TagsConst.java
@@ -1,5 +1,0 @@
-package com.corevalue.constant;
-
-public interface TagsConst {
-    String BUTTON = "button";
-}
