+++ src/test/java/com/corevalue/constant/TestConst.java
@@ -2,7 +2,6 @@ package com.corevalue.constant;
 
 public interface TestConst {
     String BASE_URL = "localhost:8085";
-//    String BASE_URL = "https://10.10.1.6:8443";
 
     int DELAY_INPUT = 200;
     int DELAY_WAIT_BUTTON = 90;
