+++ src/test/java/com/corevalue/test/xmenu/management/ClientsTest.java
@@ -7,8 +7,8 @@ import com.corevalue.constant.menu.ManagementMenuConst;
 import com.corevalue.driver.BrowserMap;
 import com.corevalue.driver.TestGroup;
 import com.corevalue.page.impl.AuthorizedLandingPage;
-import com.corevalue.submenu.impl.ManagementManageClientsSubmenu;
-import com.corevalue.submenu.impl.ManagementManageGroupsSubmenu;
+import com.corevalue.xsubmenu.impl.ManagementManageClientsSubmenu;
+import com.corevalue.xsubmenu.impl.ManagementManageGroupsSubmenu;
 import com.corevalue.test.AbstractAfterMethod;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
