+++ src/test/java/com/corevalue/test/xmenu/management/OfficesTest.java
@@ -7,8 +7,8 @@ import com.corevalue.constant.menu.ManagementMenuConst;
 import com.corevalue.driver.BrowserMap;
 import com.corevalue.driver.TestGroup;
 import com.corevalue.page.impl.AuthorizedLandingPage;
-import com.corevalue.submenu.impl.ManagementManageOfficesSubmenu;
-import com.corevalue.submenu.impl.obj.Office;
+import com.corevalue.xsubmenu.impl.ManagementManageOfficesSubmenu;
+import com.corevalue.xsubmenu.impl.obj.Office;
 import com.corevalue.test.AbstractAfterMethod;
 import org.openqa.selenium.WebDriver;
 import org.testng.Assert;
