+++ src/test/java/com/corevalue/xmenu/AbstractMenu.java
@@ -1,0 +1,13 @@
+package com.corevalue.xmenu;
+
+import com.corevalue.driver.BrowserMap;
+import com.corevalue.driver.IBrowser;
+import com.corevalue.driver.TestGroup;
+
+public interface AbstractMenu {
+    void goTo(TestGroup group);
+
+    default IBrowser browser(TestGroup group) {
+        return BrowserMap.INSTANCE.getDrivers().get(group);
+    }
+}
