+++ src/test/resources/testng.xml
@@ -1,24 +1,32 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 
 <suite name="Caseware: x-menu" verbose="1" parallel="tests" thread-count="2">
-    <test name="init: authorization and version info" >
-        <parameter name="testGroup" value="INITIAL" />
-        <classes>
-            <class name="com.corevalue.test.xmenu.initial.InitialTestSetting" />
-            <class name="com.corevalue.test.xmenu.initial.LoginTest" />
-            <class name="com.corevalue.test.xmenu.initial.HelpMenuTest"/>
-        </classes>
-    </test>
+    <!--<test name="init: authorization and version info" >-->
+        <!--<parameter name="testGroup" value="INITIAL" />-->
+        <!--<classes>-->
+            <!--<class name="com.corevalue.test.xmenu.initial.InitialTestSetting" />-->
+            <!--<class name="com.corevalue.test.xmenu.initial.LoginTest" />-->
+            <!--<class name="com.corevalue.test.xmenu.initial.HelpMenuTest"/>-->
+        <!--</classes>-->
+    <!--</test>-->
 
-    <test name="menus: Management test">
-        <parameter name="testGroup" value="MANAGEMENT" />
-        <groups>
-            <run>
-                <include name="management" />
-            </run>
-        </groups>
-        <packages>
-            <package name="com.corevalue.test.xmenu.management" />
-        </packages>
+    <!--<test name="menus: Management test">-->
+        <!--<parameter name="testGroup" value="MANAGEMENT" />-->
+        <!--<groups>-->
+            <!--<run>-->
+                <!--<include name="management" />-->
+            <!--</run>-->
+        <!--</groups>-->
+        <!--<packages>-->
+            <!--<package name="com.corevalue.test.xmenu.management" />-->
+        <!--</packages>-->
+    <!--</test>-->
+
+    <test name="init: doc manager" >
+        <parameter name="testGroup" value="DOC_MANAGER" />
+        <classes>
+            <class name="com.corevalue.test.ymenu.initial.DocManagerTestSetting" />
+            <class name="com.corevalue.test.ymenu.initial.InitTest" />
+        </classes>
     </test>
 </suite>
